initial default values small refactor
